Fixed minor bugs

+++ feature_extractors/segmentation/pixels_per_class.py
@@ -26,7 +26,7 @@ class PixelsPerClass(SegmentationFeatureExtractorAbstract):
         hist = dict.fromkeys(self._hist.keys(), 0.)
         for cls in self._hist:
             if len(self._hist[cls]):
-                hist[cls] = np.mean(self._hist[cls])
+                hist[cls] = float(np.mean(self._hist[cls]))
 
         create_bar_plot(ax, hist.values(), self._hist.keys(),
                         x_label="Class", y_label="Average # Pixels per object", title="Average Pixels Per Object",
